requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
cryptography==3.4.8
decorator==5.1.0
Django==3.2.7
django-filter==2.4.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
dynaconf==3.1.7
hvac==0.11.0
idna==3.2
knox==0.1.14
loguru==0.5.3
Markdown==3.3.4
pycparser==2.20
pytz==2021.1
requests==2.26.0
shortuuid==1.0.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.6
validators==0.18.2
